Cleaning kerberos cache files on unstage phase
Extending initial support for kerberos-based ticket mounting.
Previously, cache has been written directly to the expected file - krb5cc_*.
With this change, cache is written to the base64(volumeID) file,
and then symlink krb5cc_* created pointing to the file with the cache.
During unstage phase only volumeID and mount path is available, and
having volumeID it's possible to do the cleanup with simple traversing
directory containing the caches.